Ensure cast uses same 32-bit int precision on all platforms
Doing so allows Linux-based compilers to optimise these hot inline
functions to use smaller 32-bit instructions. Testing via callgrind
suggests the colour conversion functions are up to ~20% faster.

+++ libheif/common_utils.h
@@ -93,7 +93,7 @@ inline uint16_t clip_int_u16(int32_t x, uint16_t maxi)
 
 inline uint16_t clip_f_u16(float fx, int32_t maxi)
 {
-  long x = (long int) (fx + 0.5f);
+  int32_t x = (int32_t) (fx + 0.5f);
   if (x < 0) return 0;
   if (x > maxi) return (uint16_t) maxi;
   return static_cast<uint16_t>(x);
@@ -101,7 +101,7 @@ inline uint16_t clip_f_u16(float fx, int32_t maxi)
 
 inline uint8_t clip_f_u8(float fx)
 {
-  long x = (long int) (fx + 0.5f);
+  int32_t x = (int32_t) (fx + 0.5f);
   if (x < 0) return 0;
   if (x > 255) return 255;
   return static_cast<uint8_t>(x);
